+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Support parameterised types other than `Array` and `Object`, such as `Foo<T>`
 
 ## [0.3.5] - 2017-01-01
 - Properties are now emitted as `readonly` when applicable.